Q&A-Chatbot/requirements.txt
langchain
ipykernel
python-dotenv
langchain-community
pypdf
bs4
arxiv
pymupdf
wikipedia
arxiv
wikipedia
langchain-text-splitters
lxml
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
langchain-openai
beautifulsoup4
streamlit
langchain-ollama
